Share icons-next.json manifest validation between gen and verify
